New Algorithms
Add Functional Energy Attribute Computation for Component Trees. The first algorithm is based on a Binary Heap (heuristic false) and the other is based on the Sum of Gradient Contour (heuristic true). I also added a method remove(T elem) in PriorityQueueHeap class. The implementation of the Functional Variational Energy algorithm is under construction.