* Detailed commit:
- Added:

-> mmlib4j.representation.mergetree

A package devoted to provide a tree representation to made merges
keeping the original tree unchanged. Currently, it only works with
Component Trees.

-> InfoMergedTree

A Component Tree Merged representation. For each merge it allocates at
most one node (there are some internal optimisations to decrease
allocations). When a node is merged with its parent a new node is
allocated in InfoMergedTree. There are two building orders:

-> InfoMergedTreeLevelOrder

Must be constructed performing a Level Order Search in Component Tree.
It is useful to make filterings based on subtractive rule.

-> InfoMergedTreeReversedLevelOrder

Must be constructed performing a Reversed Level Order Search in
Component Tree. It is useful to make filterings based on direct rule.

-> AttributeUpdatedIncrementally, ComputerBasicAttributeMergeTree and
ComputerCentralMomentAttributeMergeTree

They are applied to update attributes in the InfoMergeTree.

- Modified (only major)

-> mmlib4j.descriptors.profiles

I decided to alter and move the attribute profiles helpers to here.

-> ImageFactory

Now it is an abstract class that implements the Abstract Factory design
pattern. By default it contains a field "instance" that allocates
mmlib4j images.

-> ConnectedFilteringByComponentTree

Now it supports direct and subtractive filtering strategies based on
InfoMergeTree.

- Moved

-> All unsafe packages were joined in an other Java package called
mmlib4j-unsafe.