Some Modifications
Added -> AttributeFilters (beta version)

This class is introduced to deal with attribute filters, it performs
the most common filtering rules: min, max, direct and subtractive for
morphological trees. We also intend to extend it to viterbi rule in
a future commit.

Removed ->

We moved to basic attributes computation all attributes that are not
directly related to the tree. It includes: highest and lowest x and y
values, highest and lowest x and y pixels, sum of x, and sum of y. Now
all those attributes can be accessed by method getAttributeValue passing
the right identifier (defined in Attribute.class).

Deprecated ->

I marked ConnectedFilteringByComponentTree and
ConnectedFilteringByTreeOfShape as deprecated, since I guess we can join
them in AttributeFilters.

I also marked the methods addAttributeInNodesCT and
addAttributeInNodesToS as deprecated, since now the nodes are
represented by a more general class called NodeLevelSets. I propose to
use the method addAttributeInNodes.